feat: Constant output. (#98)
Rather than open a stream per song, we'll now do the following:

1. Open a stream on startup.
2. On play, output the various audio song files to a mixer, transcoding
   if necessary.
3. On stop, cease playing those files to the mixer.

This will allow significantly faster performance as well as future
improvements that allow for things like live kick drum triggers.

+++ src/config/audio.rs
@@ -68,13 +68,13 @@ impl Audio {
         self.sample_rate.unwrap_or(44100)
     }
 
-    /// Returns the target sample format (default: Int)
+    /// Returns the target sample format (default: Float)
     pub fn sample_format(&self) -> Result<SampleFormat, Box<dyn Error>> {
         match self.sample_format.as_deref() {
             Some("float") | Some("Float") => Ok(SampleFormat::Float),
             Some("int") | Some("Int") => Ok(SampleFormat::Int),
             Some(format) => Err(format!("Unsupported sample format: {}", format).into()),
-            None => Ok(SampleFormat::Int), // Default to integer
+            None => Ok(SampleFormat::Float), // Default to float for better precision
         }
     }
